Extract tests for same to own file
